requirements.txt
avro==1.11.0
kafka-python==2.0.2